#11: Reworked BackendConfigUi class hierarchy
