+++ src/main/scala/org/fs/mael/core/backend/BackendConfigUiImpl.scala
@@ -7,45 +7,20 @@ import org.eclipse.swt.widgets.Composite
 import org.eclipse.swt.widgets.TabFolder
 import org.eclipse.swt.widgets.TabItem
 import org.fs.mael.core.config.ConfigManager
-import org.fs.mael.core.config.InMemoryConfigManager
-import org.fs.mael.core.utils.CoreUtils._
 import org.fs.mael.ui.config.MFieldEditorPreferencePage
 import org.fs.mael.ui.config.MPreferencePageDescriptor
 import org.fs.mael.ui.utils.SwtUtils
 
 class BackendConfigUiImpl(
-  backendId:  String,
-  pageDescr:  Seq[MPreferencePageDescriptor[_ <: MFieldEditorPreferencePage]],
-  isEditable: Boolean
-) extends BackendConfigUi {
+  override val backendId:        String,
+  override val isEditable:       Boolean,
+  override val cfgOption:        Option[ConfigManager],
+  override val globalCfgMgr:     ConfigManager,
+  override val tabFolder:        TabFolder,
+  override val pageDescriptions: Seq[MPreferencePageDescriptor[_ <: MFieldEditorPreferencePage]]
+) extends AbstractBackendConfigUi {
 
-  val cfgMgr = new InMemoryConfigManager
-
-  var pages: Seq[MFieldEditorPreferencePage] = Seq.empty
-
-  def initialize(
-    cfgOption:    Option[ConfigManager],
-    tabFolder:    TabFolder,
-    globalCfgMgr: ConfigManager
-  ): Unit = {
-    cfgOption match {
-      case Some(cfg) => cfgMgr.resetTo(cfg, backendId) // Ignore default preferences
-      case None      => cfgMgr.resetTo(globalCfgMgr, backendId)
-    }
-    pageDescr.foreach { pageDef =>
-      createPage(pageDef, tabFolder)
-    }
-  }
-
-  override def get(): InMemoryConfigManager = {
-    require(!pages.isEmpty, "Forgot to call initialize()?")
-    if (isEditable) {
-      requireFriendly(pages.forall(_.performOk), "Some settings are invalid")
-    }
-    cfgMgr
-  }
-
-  private def createPage(pageDescr: MPreferencePageDescriptor[_ <: MFieldEditorPreferencePage], tabFolder: TabFolder): Unit = {
+  override def createPage[T <: MFieldEditorPreferencePage](pageDescr: MPreferencePageDescriptor[T], tabFolder: TabFolder): T = {
     val tab = new TabItem(tabFolder, SWT.NONE)
     tab.setText(pageDescr.name)
     val container = new Composite(tabFolder, SWT.NONE)
@@ -62,7 +37,6 @@ class BackendConfigUiImpl(
         case (editor, parent) => SwtUtils.disable(editor, parent)
       }
     }
-
-    pages = pages :+ page
+    page
   }
 }
